Update version of library, Omise PHP, to 2.7.1
The objective for update version of the library, Omise PHP, is the new
version has been added a user agent that indicated the version of PHP
that run on the merchant's server.

The user agent will be sent to the Omise API when crate a Omise charge
and it will be stored in the log at the Omise API.

The user agent log can be found at the Omise dashboard.