Change a front-end message, please select IB
Before change, the message displayed to the customer that the customer
need to select any `provider` from the list.

After change, the word `provider` has been changed to `bank`.

The reason for this change is at the front-end, internet banking form,
it display or it has only a word `bank`. Such as `Internet Banking`,
`Siam Commercial Bank` and etc. If the message display that the
customer need to select `provider`, it can be confused.

So, the word `provider` has been changed to `bank`.